Text indexing added - Now with less retardation.
Just kidding it is just retarded in new ways.
It replaces non alpha characters with spaces and then discards
words with less than two characters.  doesn't -> doesn t -> doesn

We need a stemmer it would sort this out if we could remove punctation that
bordered tokens and then stem them and then store them however i think anything
we do will be imperfect.  It pretty much works and it is tested that it is
building this secondary db.  Its pretty fucking wild.